requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pillow
numpy>=1.22,<1.24
python-multipart
torch==2.1.0
torchvision==0.16.0
timm==0.9.7